Extract TestAccessControl
